// src/app/core/models/requests/entryOrder.model.ts
export interface IEntryOrder {
  unitadministrative: string; // Unidad Administrativa
  orderDate: string; // Fecha de Orden
  concept: string; // Concepto
  paymentMethod: string; // Forma de Pago
  amount: number; // Estado
  referenceNo: string; // Numero de Referencia
  bank: string; // Institucion Bancaria
}